chore(mail): move template to util
